Move preferences-api to @theia/core
By keeping the actual preference providers in another package but the
api in core, this allows core components to also make use of the
preferences.
