+++ .vscode/settings.json
@@ -24,7 +24,6 @@
         "packages/monaco/coverage/lcov.info",
         "packages/navigator/coverage/lcov.info",
         "packages/keymaps/coverage/lcov.info",
-        "packages/preferences-api/coverage/lcov.info",
         "packages/preferences/coverage/lcov.info",
         "packages/process/coverage/lcov.info",
         "packages/python/coverage/lcov.info",
