+++ packages/core/package.json
@@ -11,6 +11,7 @@
     "@types/express": "^4.0.36",
     "@types/ws": "^3.0.2",
     "@types/yargs": "^8.0.2",
+    "ajv": "^5.2.2",
     "body-parser": "^1.17.2",
     "bunyan": "^1.8.10",
     "electron": "1.7.11",
