+++ packages/core/src/browser/frontend-application-module.ts
@@ -32,6 +32,7 @@ import {
 import { StatusBar, StatusBarImpl } from "./status-bar/status-bar";
 import { LabelParser } from './label-parser';
 import { LabelProvider, LabelProviderContribution, DefaultUriLabelProviderContribution } from "./label-provider";
+import { PreferenceService, PreferenceServiceImpl, PreferenceProviders } from './preferences';
 
 import '../../src/browser/style/index.css';
 import 'font-awesome/css/font-awesome.min.css';
@@ -108,6 +109,12 @@ export const frontendApplicationModule = new ContainerModule((bind, unbind, isBo
     bind(LabelProviderContribution).to(DefaultUriLabelProviderContribution).inSingletonScope();
 
     bind(CommandContribution).to(ThemingCommandContribution).inSingletonScope();
+
+    bind(PreferenceProviders).toFactory(ctx => () => []);
+    bind(PreferenceServiceImpl).toSelf().inSingletonScope();
+    for (const serviceIdentifier of [PreferenceService, FrontendApplicationContribution]) {
+        bind(serviceIdentifier).toDynamicValue(ctx => ctx.container.get(PreferenceServiceImpl)).inSingletonScope();
+    }
 });
 
 const theme = ThemeService.get().getCurrentTheme().id;
