+++ packages/core/src/browser/preferences/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 TypeFox and others.
+ * Copyright (C) 2017 TypeFox and others.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
