+++ packages/git/package.json
@@ -6,7 +6,6 @@
     "@theia/core": "^0.3.3",
     "@theia/editor": "^0.3.4",
     "@theia/filesystem": "^0.3.4",
-    "@theia/preferences-api": "^0.3.3",
     "@theia/workspace": "^0.3.4",
     "@types/fs-extra": "^4.0.2",
     "dugite-extra": "0.0.1-alpha.15",
