+++ packages/monaco/package.json
@@ -9,7 +9,6 @@
     "@theia/languages": "^0.3.3",
     "@theia/markers": "^0.3.4",
     "@theia/outline-view": "^0.3.4",
-    "@theia/preferences-api": "^0.3.3",
     "@theia/workspace": "^0.3.4",
     "monaco-css": "^1.3.3",
     "monaco-html": "^1.3.2",
