--- packages/preferences-api/README.md
@@ -1,6 +1,0 @@
-# Theia - Preferences API
-
-See [here](https://github.com/theia-ide/theia) for a detailed documentation.
-
-## License
-[Apache-2.0](https://github.com/theia-ide/theia/blob/master/LICENSE)
