--- packages/preferences-api/compile.tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../configs/base.tsconfig",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "lib"
-  },
-  "include": [
-    "src"
-  ]
-}
