--- packages/preferences-api/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@theia/preferences-api",
-  "version": "0.3.3",
-  "description": "Theia - Preferences API",
-  "main": "lib/browser/index.js",
-  "typings": "lib/browser/index.d.ts",
-  "dependencies": {
-    "@theia/core": "^0.3.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "keywords": [
-    "theia-extension"
-  ],
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/theia-ide/theia.git"
-  },
-  "bugs": {
-    "url": "https://github.com/theia-ide/theia/issues"
-  },
-  "homepage": "https://github.com/theia-ide/theia",
-  "files": [
-    "lib",
-    "src"
-  ],
-  "scripts": {
-    "prepare": "yarn run clean && yarn run build",
-    "clean": "theiaext clean",
-    "build": "theiaext build",
-    "watch": "theiaext watch",
-    "test": "theiaext test",
-    "docs": "theiaext docs"
-  },
-  "devDependencies": {
-    "@theia/ext-scripts": "^0.2.0"
-  },
-  "nyc": {
-    "extends": "../../configs/nyc.json"
-  }
-}
