+++ packages/preferences/package.json
@@ -5,11 +5,9 @@
   "dependencies": {
     "@theia/core": "^0.3.3",
     "@theia/filesystem": "^0.3.4",
-    "@theia/preferences-api": "^0.3.3",
     "@theia/userstorage": "^0.3.4",
     "@theia/workspace": "^0.3.4",
     "@types/fs-extra": "^4.0.2",
-    "ajv": "^5.2.2",
     "fs-extra": "^4.0.2",
     "jsonc-parser": "^1.0.0"
   },
