+++ packages/userstorage/src/browser/user-storage-service-filesystem.spec.ts
@@ -16,8 +16,8 @@ import { MockLogger } from '@theia/core/lib/common/test/mock-logger';
 import { FileSystem, FileStat } from '@theia/filesystem/lib/common/';
 import { FileSystemPreferences, createFileSystemPreferences } from '@theia/filesystem/lib/browser/filesystem-preferences';
 import { FileSystemWatcher, FileChange, FileChangeType } from '@theia/filesystem/lib/browser/filesystem-watcher';
-import { PreferenceService } from '@theia/preferences-api/lib/browser';
-import { MockPreferenceService } from '@theia/preferences-api/lib/browser/test/mock-preference-service';
+import { PreferenceService } from "@theia/core/lib/browser/preferences";
+import { MockPreferenceService } from '@theia/core/lib/browser/preferences/test/mock-preference-service';
 import { FileSystemWatcherServer } from '@theia/filesystem/lib/common/filesystem-watcher-protocol';
 import { MockFilesystem, MockFilesystemWatcherServer } from '@theia/filesystem/lib/common/test';
 import { UserStorageUri } from './user-storage-uri';
